Add and divide Interface module
