+++ src/modules/battle.js
@@ -1,0 +1,178 @@
+export const BattleScene = new Phaser.Class({
+
+  Extends: Phaser.Scene,
+
+  initialize:
+
+  function BattleScene ()
+  {
+      Phaser.Scene.call(this, { key: "BattleScene" });
+  },
+  create: function ()
+  {    
+      // change the background to green
+      this.cameras.main.setBackgroundColor("rgba(0, 200, 0, 0.5)");
+      this.startBattle();
+      // on wake event we call startBattle too
+      this.sys.events.on('wake', this.startBattle, this);             
+  },
+  startBattle: function() {
+      // player character - warrior
+      const warrior = new PlayerCharacter(this, 250, 50, "player", 1, "Warrior", 100, 20);        
+      this.add.existing(warrior);
+      
+      // player character - mage
+      const mage = new PlayerCharacter(this, 250, 100, "player", 4, "Mage", 80, 8);
+      this.add.existing(mage);            
+      
+      const dragonblue = new Enemy(this, 50, 50, "dragonblue", null, "Dragon", 50, 3);
+      this.add.existing(dragonblue);
+      
+      const dragonOrange = new Enemy(this, 50, 100, "dragonorrange", null,"Dragon2", 50, 3);
+      this.add.existing(dragonOrange);
+      
+      // array with heroes
+      this.heroes = [ warrior, mage ];
+      // array with enemies
+      this.enemies = [ dragonblue, dragonOrange ];
+      // array with both parties, who will attack
+      this.units = this.heroes.concat(this.enemies);
+      
+      this.index = -1; // currently active unit
+      
+      this.scene.run("UIScene");        
+  },
+  nextTurn: function() {  
+      // if we have victory or game over
+      if(this.checkEndBattle()) {           
+          this.endBattle();
+          return;
+      }
+      do {
+          // currently active unit
+          this.index++;
+          // if there are no more units, we start again from the first one
+          if(this.index >= this.units.length) {
+              this.index = 0;
+          }            
+      } while(!this.units[this.index].living);
+      // if its player hero
+      if(this.units[this.index] instanceof PlayerCharacter) {
+          // we need the player to select action and then enemy
+          this.events.emit("PlayerSelect", this.index);
+      } else { // else if its enemy unit
+          // pick random living hero to be attacked
+          let r;
+          do {
+              r = Math.floor(Math.random() * this.heroes.length);
+          } while(!this.heroes[r].living) 
+          // call the enemy's attack function 
+          this.units[this.index].attack(this.heroes[r]);  
+          // add timer for the next turn, so will have smooth gameplay
+          this.time.addEvent({ delay: 1500, callback: this.nextTurn, callbackScope: this });
+      }
+  },     
+  // check for game over or victory
+  checkEndBattle: function() {        
+      let victory = true;
+      // if all enemies are dead we have victory
+      for(let i = 0; i < this.enemies.length; i++) {
+          if(this.enemies[i].living)
+              victory = false;
+      }
+      let gameOver = true;
+      // if all heroes are dead we have game over
+      for(let i = 0; i < this.heroes.length; i++) {
+          if(this.heroes[i].living)
+              gameOver = false;
+      }
+      return victory || gameOver;
+  },
+  // when the player have selected the enemy to be attacked
+  receivePlayerSelection: function(action, target) {
+      if(action == "attack") {            
+          this.units[this.index].attack(this.enemies[target]);              
+      }
+      // next turn in 1.7 seconds
+      this.time.addEvent({ delay: 1700, callback: this.nextTurn, callbackScope: this });        
+  },    
+  endBattle: function() {       
+      // clear state, remove sprites
+      this.heroes.length = 0;
+      this.enemies.length = 0;
+      for(let i = 0; i < this.units.length; i++) {
+          // link item
+          this.units[i].destroy();            
+      }
+      this.units.length = 0;
+      // sleep the UI
+      this.scene.sleep('UIScene');
+      // return to WorldScene and sleep current BattleScene
+      this.scene.switch('WorldScene');
+  }
+});
+
+// base class for heroes and enemies
+export const Unit = new Phaser.Class({
+  Extends: Phaser.GameObjects.Sprite,
+
+  initialize:
+
+  function Unit(scene, x, y, texture, frame, type, hp, damage) {
+      Phaser.GameObjects.Sprite.call(this, scene, x, y, texture, frame)
+      this.type = type;
+      this.maxHp = this.hp = hp;
+      this.damage = damage; // default damage     
+      this.living = true;         
+      this.menuItem = null;
+  },
+  // we will use this to notify the menu item when the unit is dead
+  setMenuItem: function(item) {
+      this.menuItem = item;
+  },
+  // attack the target unit
+  attack: function(target) {
+      if(target.living) {
+          target.takeDamage(this.damage);
+          this.scene.events.emit("Message", this.type + " attacks " + target.type + " for " + this.damage + " damage");
+      }
+  },    
+  takeDamage: function(damage) {
+      this.hp -= damage;
+      if(this.hp <= 0) {
+          this.hp = 0;
+          this.menuItem.unitKilled();
+          this.living = false;
+          this.visible = false;   
+          this.menuItem = null;
+      }
+  }    
+});
+
+export const Enemy = new Phaser.Class({
+  Extends: Unit,
+
+  initialize:
+  function Enemy(scene, x, y, texture, frame, type, hp, damage) {
+      Unit.call(this, scene, x, y, texture, frame, type, hp, damage);
+      this.setScale(1); // Size of the enemy
+  }
+});
+
+const PlayerCharacter = new Phaser.Class({
+  Extends: Unit,
+
+  initialize:
+  function PlayerCharacter(scene, x, y, texture, frame, type, hp, damage) {
+      Unit.call(this, scene, x, y, texture, frame, type, hp, damage);
+      // flip the image so I don"t have to edit it manually
+      this.flipX = true;
+      
+      this.setScale(1.5);
+  }
+});
+
+
+
+
+
